feat: refacto report page to externalize in component and use it with vehicle
